+++ dev-site-config/generatedComponentConfig.js
@@ -20,16 +20,17 @@ import TerraSiteIndex14 from '../packages/terra-site/src/examples/date-time-pick
 import TerraSiteIndex15 from '../packages/terra-site/src/examples/demographics-banner/Index';
 import TerraSiteIndex16 from '../packages/terra-site/src/examples/dialog/Index';
 import TerraSiteIndex17 from '../packages/terra-site/src/examples/divider/Index';
-import TerraSiteIndex18 from '../packages/terra-site/src/examples/dynamic-grid/Index';
-import TerraSiteIndex19 from '../packages/terra-site/src/examples/embedded-content-consumer/Index';
-import TerraSiteIndex20 from '../packages/terra-site/src/examples/form-checkbox/Index';
-import TerraSiteIndex21 from '../packages/terra-site/src/examples/form-field/Index';
-import TerraSiteIndex22 from '../packages/terra-site/src/examples/form-fieldset/Index';
-import TerraSiteIndex23 from '../packages/terra-site/src/examples/form-input/Index';
-import TerraSiteIndex24 from '../packages/terra-site/src/examples/form-radio/Index';
-import TerraSiteIndex25 from '../packages/terra-site/src/examples/form-select/Index';
-import TerraSiteIndex26 from '../packages/terra-site/src/examples/form-textarea/Index';
-import TerraSiteIndex27 from '../packages/terra-site/src/examples/form/Index';
+import TerraSiteIndex18 from '../packages/terra-site/src/examples/doc-template/Index';
+import TerraSiteIndex19 from '../packages/terra-site/src/examples/dynamic-grid/Index';
+import TerraSiteIndex20 from '../packages/terra-site/src/examples/embedded-content-consumer/Index';
+import TerraSiteIndex21 from '../packages/terra-site/src/examples/form-checkbox/Index';
+import TerraSiteIndex22 from '../packages/terra-site/src/examples/form-field/Index';
+import TerraSiteIndex23 from '../packages/terra-site/src/examples/form-fieldset/Index';
+import TerraSiteIndex24 from '../packages/terra-site/src/examples/form-input/Index';
+import TerraSiteIndex25 from '../packages/terra-site/src/examples/form-radio/Index';
+import TerraSiteIndex26 from '../packages/terra-site/src/examples/form-select/Index';
+import TerraSiteIndex27 from '../packages/terra-site/src/examples/form-textarea/Index';
+import TerraSiteIndex28 from '../packages/terra-site/src/examples/form/Index';
 import TerraSiteControlIndex from '../packages/terra-site/src/examples/form/components/ControlIndex';
 import TerraSiteFieldIndex from '../packages/terra-site/src/examples/form/components/FieldIndex';
 import TerraSiteFieldsetIndex from '../packages/terra-site/src/examples/form/components/FieldsetIndex';
@@ -40,38 +41,38 @@ import TerraSiteSelectIndex from '../packages/terra-site/src/examples/form/compo
 import TerraSiteTextFieldIndex from '../packages/terra-site/src/examples/form/components/TextFieldIndex';
 import TerraSiteTextareaFieldIndex from '../packages/terra-site/src/examples/form/components/TextareaFieldIndex';
 import TerraSiteTextareaIndex from '../packages/terra-site/src/examples/form/components/TextareaIndex';
-import TerraSiteIndex28 from '../packages/terra-site/src/examples/grid/Index';
-import TerraSiteIndex29 from '../packages/terra-site/src/examples/heading/Index';
-import TerraSiteIndex30 from '../packages/terra-site/src/examples/i18n/Index';
-import TerraSiteIndex31 from '../packages/terra-site/src/examples/icon/Index';
+import TerraSiteIndex29 from '../packages/terra-site/src/examples/grid/Index';
+import TerraSiteIndex30 from '../packages/terra-site/src/examples/heading/Index';
+import TerraSiteIndex31 from '../packages/terra-site/src/examples/i18n/Index';
+import TerraSiteIndex32 from '../packages/terra-site/src/examples/icon/Index';
 import TerraSiteIconStatic from '../packages/terra-site/src/examples/icon/components/IconStatic';
 import TerraSiteIconThemeable from '../packages/terra-site/src/examples/icon/components/IconThemeable';
-import TerraSiteIndex32 from '../packages/terra-site/src/examples/image/Index';
-import TerraSiteIndex33 from '../packages/terra-site/src/examples/list/Index';
+import TerraSiteIndex33 from '../packages/terra-site/src/examples/image/Index';
+import TerraSiteIndex34 from '../packages/terra-site/src/examples/list/Index';
 import TerraSiteMultiSelectIndex from '../packages/terra-site/src/examples/list/components/MultiSelectIndex';
 import TerraSiteSelectableListIndex from '../packages/terra-site/src/examples/list/components/SelectableListIndex';
 import TerraSiteSingleSelectIndex from '../packages/terra-site/src/examples/list/components/SingleSelectIndex';
-import TerraSiteIndex34 from '../packages/terra-site/src/examples/menu/Index';
-import TerraSiteIndex35 from '../packages/terra-site/src/examples/overlay/Index';
-import TerraSiteIndex36 from '../packages/terra-site/src/examples/profile-image/Index';
-import TerraSiteIndex37 from '../packages/terra-site/src/examples/progress-bar/Index';
-import TerraSiteIndex38 from '../packages/terra-site/src/examples/responsive-element/Index';
-import TerraSiteIndex39 from '../packages/terra-site/src/examples/search-field/Index';
-import TerraSiteIndex40 from '../packages/terra-site/src/examples/section-header/Index';
-import TerraSiteIndex41 from '../packages/terra-site/src/examples/signature/Index';
-import TerraSiteIndex42 from '../packages/terra-site/src/examples/slide-group/Index';
-import TerraSiteIndex43 from '../packages/terra-site/src/examples/spacer/Index';
-import TerraSiteIndex44 from '../packages/terra-site/src/examples/status-view/Index';
-import TerraSiteIndex45 from '../packages/terra-site/src/examples/status/Index';
-import TerraSiteIndex46 from '../packages/terra-site/src/examples/table/Index';
+import TerraSiteIndex35 from '../packages/terra-site/src/examples/menu/Index';
+import TerraSiteIndex36 from '../packages/terra-site/src/examples/overlay/Index';
+import TerraSiteIndex37 from '../packages/terra-site/src/examples/profile-image/Index';
+import TerraSiteIndex38 from '../packages/terra-site/src/examples/progress-bar/Index';
+import TerraSiteIndex39 from '../packages/terra-site/src/examples/responsive-element/Index';
+import TerraSiteIndex40 from '../packages/terra-site/src/examples/search-field/Index';
+import TerraSiteIndex41 from '../packages/terra-site/src/examples/section-header/Index';
+import TerraSiteIndex42 from '../packages/terra-site/src/examples/signature/Index';
+import TerraSiteIndex43 from '../packages/terra-site/src/examples/slide-group/Index';
+import TerraSiteIndex44 from '../packages/terra-site/src/examples/spacer/Index';
+import TerraSiteIndex45 from '../packages/terra-site/src/examples/status-view/Index';
+import TerraSiteIndex46 from '../packages/terra-site/src/examples/status/Index';
+import TerraSiteIndex47 from '../packages/terra-site/src/examples/table/Index';
 import TerraSiteMultiSelectIndex1 from '../packages/terra-site/src/examples/table/components/MultiSelectIndex';
-import TerraSiteIndex47 from '../packages/terra-site/src/examples/tabs/Index';
-import TerraSiteIndex48 from '../packages/terra-site/src/examples/tag/Index';
-import TerraSiteIndex49 from '../packages/terra-site/src/examples/text/Index';
-import TerraSiteIndex50 from '../packages/terra-site/src/examples/time-input/Index';
-import TerraSiteIndex51 from '../packages/terra-site/src/examples/toggle-button/Index';
-import TerraSiteIndex52 from '../packages/terra-site/src/examples/toggle-section-header/Index';
-import TerraSiteIndex53 from '../packages/terra-site/src/examples/toggle/Index';
+import TerraSiteIndex48 from '../packages/terra-site/src/examples/tabs/Index';
+import TerraSiteIndex49 from '../packages/terra-site/src/examples/tag/Index';
+import TerraSiteIndex50 from '../packages/terra-site/src/examples/text/Index';
+import TerraSiteIndex51 from '../packages/terra-site/src/examples/time-input/Index';
+import TerraSiteIndex52 from '../packages/terra-site/src/examples/toggle-button/Index';
+import TerraSiteIndex53 from '../packages/terra-site/src/examples/toggle-section-header/Index';
+import TerraSiteIndex54 from '../packages/terra-site/src/examples/toggle/Index';
 
 // Component Test Examples
 import AbstractModalAppendClass from '../packages/terra-abstract-modal/tests/nightwatch/components/AbstractModalAppendClass';
@@ -197,6 +198,7 @@ import PopulatedDemographicsBanner from '../packages/terra-demographics-banner/t
 import PostMenstrualAgeDemographicsBannerNoLabel from '../packages/terra-demographics-banner/tests/nightwatch/PostMenstrualAgeDemographicsBannerNoLabel';
 import DefaultDialog from '../packages/terra-dialog/tests/nightwatch/DefaultDialog';
 import DefaultDivider from '../packages/terra-divider/tests/nightwatch/DefaultDivider';
+import DefaultDocTemplate from '../packages/terra-doc-template/examples/test-examples/DefaultDocTemplate.example';
 import DefaultDynamicGrid from '../packages/terra-dynamic-grid/tests/nightwatch/DefaultDynamicGrid';
 import LargeDynamicGrid from '../packages/terra-dynamic-grid/tests/nightwatch/LargeDynamicGrid';
 import BasicConsumer from '../packages/terra-embedded-content-consumer/tests/nightwatch/consumers/BasicConsumer';
@@ -1250,6 +1252,17 @@ const componentConfig = {
         name: 'Default Divider',
         path: '/default-divider',
         component: DefaultDivider,
+      },
+    ],
+  },
+  'terra-doc-template': {
+    name: 'Doc Template',
+    path: '/doc-template',
+    tests: [
+      {
+        name: 'Default Doc Template',
+        path: '/default-doc-template',
+        component: DefaultDocTemplate,
       },
     ],
   },
@@ -2777,13 +2790,24 @@ const componentConfig = {
         ],
       },
       {
+        name: 'Doc Template',
+        path: '/doc-template',
+        pages: [
+          {
+            name: 'Index',
+            path: '/index',
+            component: TerraSiteIndex18,
+          },
+        ],
+      },
+      {
         name: 'Dynamic Grid',
         path: '/dynamic-grid',
         pages: [
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex18,
+            component: TerraSiteIndex19,
           },
         ],
       },
@@ -2794,7 +2818,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex19,
+            component: TerraSiteIndex20,
           },
         ],
       },
@@ -2805,7 +2829,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex20,
+            component: TerraSiteIndex21,
           },
         ],
       },
@@ -2816,7 +2840,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex21,
+            component: TerraSiteIndex22,
           },
         ],
       },
@@ -2827,7 +2851,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex22,
+            component: TerraSiteIndex23,
           },
         ],
       },
@@ -2838,7 +2862,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex23,
+            component: TerraSiteIndex24,
           },
         ],
       },
@@ -2849,7 +2873,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex24,
+            component: TerraSiteIndex25,
           },
         ],
       },
@@ -2860,7 +2884,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex25,
+            component: TerraSiteIndex26,
           },
         ],
       },
@@ -2871,7 +2895,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex26,
+            component: TerraSiteIndex27,
           },
         ],
       },
@@ -2882,7 +2906,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex27,
+            component: TerraSiteIndex28,
           },
           {
             name: 'Control Index',
@@ -2943,7 +2967,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex28,
+            component: TerraSiteIndex29,
           },
         ],
       },
@@ -2954,7 +2978,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex29,
+            component: TerraSiteIndex30,
           },
         ],
       },
@@ -2965,7 +2989,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex30,
+            component: TerraSiteIndex31,
           },
         ],
       },
@@ -2976,7 +3000,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex31,
+            component: TerraSiteIndex32,
           },
           {
             name: 'Icon Static',
@@ -2997,7 +3021,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex32,
+            component: TerraSiteIndex33,
           },
         ],
       },
@@ -3008,7 +3032,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex33,
+            component: TerraSiteIndex34,
           },
           {
             name: 'Multi Select Index',
@@ -3034,7 +3058,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex34,
+            component: TerraSiteIndex35,
           },
         ],
       },
@@ -3045,7 +3069,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex35,
+            component: TerraSiteIndex36,
           },
         ],
       },
@@ -3056,7 +3080,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex36,
+            component: TerraSiteIndex37,
           },
         ],
       },
@@ -3067,7 +3091,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex37,
+            component: TerraSiteIndex38,
           },
         ],
       },
@@ -3078,7 +3102,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex38,
+            component: TerraSiteIndex39,
           },
         ],
       },
@@ -3089,7 +3113,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex39,
+            component: TerraSiteIndex40,
           },
         ],
       },
@@ -3100,7 +3124,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex40,
+            component: TerraSiteIndex41,
           },
         ],
       },
@@ -3111,7 +3135,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex41,
+            component: TerraSiteIndex42,
           },
         ],
       },
@@ -3122,7 +3146,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex42,
+            component: TerraSiteIndex43,
           },
         ],
       },
@@ -3133,7 +3157,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex43,
+            component: TerraSiteIndex44,
           },
         ],
       },
@@ -3144,7 +3168,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex44,
+            component: TerraSiteIndex45,
           },
         ],
       },
@@ -3155,7 +3179,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex45,
+            component: TerraSiteIndex46,
           },
         ],
       },
@@ -3166,7 +3190,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex46,
+            component: TerraSiteIndex47,
           },
           {
             name: 'Multi Select Index',
@@ -3182,7 +3206,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex47,
+            component: TerraSiteIndex48,
           },
         ],
       },
@@ -3193,7 +3217,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex48,
+            component: TerraSiteIndex49,
           },
         ],
       },
@@ -3204,7 +3228,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex49,
+            component: TerraSiteIndex50,
           },
         ],
       },
@@ -3215,7 +3239,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex50,
+            component: TerraSiteIndex51,
           },
         ],
       },
@@ -3226,7 +3250,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex51,
+            component: TerraSiteIndex52,
           },
         ],
       },
@@ -3237,7 +3261,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex52,
+            component: TerraSiteIndex53,
           },
         ],
       },
@@ -3248,7 +3272,7 @@ const componentConfig = {
           {
             name: 'Index',
             path: '/index',
-            component: TerraSiteIndex53,
+            component: TerraSiteIndex54,
           },
         ],
       },